# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Telecom Service & Network Management System",
    "description": "Backend platform for managing telecom customers, plans, SIMs, devices, subscriptions, usage, network infrastructure, outages, support tickets, SLAs and operational analytics.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/towers": {
      "post": {
        "tags": [
          "Network Towers"
        ],
        "summary": "Create Tower",
        "operationId": "create_tower_api_v1_towers_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TowerCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TowerOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Network Towers"
        ],
        "summary": "List Towers",
        "operationId": "list_towers_api_v1_towers_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tower_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TowerStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tower Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number, starting at 1",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number, starting at 1"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Items per page"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Column name to sort by",
              "title": "Sort By"
            },
            "description": "Column name to sort by"
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "default": "asc",
              "title": "Sort Order"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Free-text search",
              "title": "Search"
            },
            "description": "Free-text search"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_TowerOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/towers/{tower_id}": {
      "get": {
        "tags": [
          "Network Towers"
        ],
        "summary": "Get Tower",
        "operationId": "get_tower_api_v1_towers__tower_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tower_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tower Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TowerOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Network Towers"
        ],
        "summary": "Update Tower",
        "operationId": "update_tower_api_v1_towers__tower_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tower_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tower Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TowerUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TowerOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/towers/{tower_id}/status": {
      "patch": {
        "tags": [
          "Network Towers"
        ],
        "summary": "Set Tower Status",
        "operationId": "set_tower_status_api_v1_towers__tower_id__status_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tower_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tower Id"
            }
          },
          {
            "name": "tower_status",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TowerStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TowerOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_TowerOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TowerOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "pages": {
            "type": "integer",
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "pages"
        ],
        "title": "PaginatedResponse[TowerOut]"
      },
...
      "TowerCreate": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tower_type": {
            "$ref": "#/components/schemas/TowerType",
            "default": "macro"
          },
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          },
          "coverage_radius_km": {
            "type": "number",
            "title": "Coverage Radius Km",
            "default": 5.0
          },
          "capacity": {
            "type": "integer",
            "title": "Capacity",
            "default": 1000
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "latitude",
          "longitude"
        ],
        "title": "TowerCreate"
      },
      "TowerOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tower_type": {
            "$ref": "#/components/schemas/TowerType"
          },
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          },
          "coverage_radius_km": {
            "type": "number",
            "title": "Coverage Radius Km"
          },
          "capacity": {
            "type": "integer",
            "title": "Capacity"
          },
          "status": {
            "$ref": "#/components/schemas/TowerStatus"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "tower_type",
          "latitude",
          "longitude",
          "coverage_radius_km",
          "capacity",
          "status"
        ],
        "title": "TowerOut"
      },
      "TowerStatus": {
        "type": "string",
        "enum": [
          "active",
          "maintenance",
          "offline",
          "decommissioned"
        ],
        "title": "TowerStatus"
      },
      "TowerType": {
        "type": "string",
        "enum": [
          "macro",
          "micro",
          "pico",
          "femto"
        ],
        "title": "TowerType"
      },
      "TowerUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "coverage_radius_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage Radius Km"
          },
          "capacity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TowerStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "TowerUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}